# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Warehouse Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  master: 'master',
  user: 'user',
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResult {
  token: string;
  user: User;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  master: 'master',
  user: 'user',
} as const;

export interface UserInput {
  username: string;
  password: string;
  role: UserInputRole;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  master: 'master',
  user: 'user',
} as const;

export interface UserUpdate {
  username?: string;
  /** @nullable */
  password?: string | null;
  role?: UserUpdateRole;
}

export type MaterialKategori = typeof MaterialKategori[keyof typeof MaterialKategori];


export const MaterialKategori = {
  scan: 'scan',
  'non-scan': 'non-scan',
} as const;

export interface Material {
  id: number;
  name: string;
  code: string;
  /** @nullable */
  description?: string | null;
  kategori: MaterialKategori;
  createdAt: string;
}

export type MaterialInputKategori = typeof MaterialInputKategori[keyof typeof MaterialInputKategori];


export const MaterialInputKategori = {
  scan: 'scan',
  'non-scan': 'non-scan',
} as const;

export interface MaterialInput {
  name: string;
  code: string;
  description?: string;
  kategori?: MaterialInputKategori;
}

export type MaterialUpdateKategori = typeof MaterialUpdateKategori[keyof typeof MaterialUpdateKategori];


export const MaterialUpdateKategori = {
  scan: 'scan',
  'non-scan': 'non-scan',
} as const;

export interface MaterialUpdate {
  name?: string;
  code?: string;
  /** @nullable */
  description?: string | null;
  kategori?: MaterialUpdateKategori;
}

export type ScanInStatus = typeof ScanInStatus[keyof typeof ScanInStatus];


export const ScanInStatus = {
  scanning: 'scanning',
  completed: 'completed',
} as const;

export interface ScanItem {
  id: number;
  serialNumber: string;
  /** @nullable */
  scanInId?: number | null;
  /** @nullable */
  scanOutId?: number | null;
  createdAt: string;
}

export interface ScanIn {
  id: number;
  materialId: number;
  materialName: string;
  boxLabel: string;
  status: ScanInStatus;
  userId: number;
  userName: string;
  /** @nullable */
  qrCodeData?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
  items: ScanItem[];
}

export interface ScanInInput {
  materialId: number;
  userId: number;
}

export type ScanInUpdateStatus = typeof ScanInUpdateStatus[keyof typeof ScanInUpdateStatus];


export const ScanInUpdateStatus = {
  scanning: 'scanning',
  completed: 'completed',
} as const;

export interface ScanInUpdate {
  status?: ScanInUpdateStatus;
  qrCodeData?: string;
}

export interface ScanOut {
  id: number;
  userId: number;
  userName: string;
  createdAt: string;
  items: ScanItem[];
}

export interface ScanOutInput {
  userId: number;
}

export interface ScanOutItemInput {
  qrData: string;
}

export interface ScanItemInput {
  /**
     * @minLength 28
     * @maxLength 28
     */
  serialNumber: string;
}

export interface ScanOutItemsBulkInput {
  serialNumbers: string[];
}

export interface ScanOutItemsBulkResult {
  scannedCount: number;
  alreadyOut: string[];
  notFound: string[];
}

export type HistoryRecordType = typeof HistoryRecordType[keyof typeof HistoryRecordType];


export const HistoryRecordType = {
  in: 'in',
  out: 'out',
} as const;

export type HistoryRecordSource = typeof HistoryRecordSource[keyof typeof HistoryRecordSource];


export const HistoryRecordSource = {
  scan: 'scan',
  'non-scan': 'non-scan',
} as const;

export interface HistoryRecord {
  id: number;
  type: HistoryRecordType;
  source: HistoryRecordSource;
  /** @nullable */
  materialId: number | null;
  /** @nullable */
  materialCode?: string | null;
  /** @nullable */
  materialName: string | null;
  /** @nullable */
  boxLabel: string | null;
  userId: number;
  userName: string;
  serialNumbers: string[];
  count: number;
  satuan?: string;
  createdAt: string;
}

export interface DashboardSummary {
  totalMaterialIn: number;
  totalMaterialOut: number;
  totalStock: number;
  totalMaterials: number;
  totalUsers: number;
}

export interface MaterialStat {
  materialId: number;
  materialCode: string;
  materialName: string;
  totalIn: number;
  totalOut: number;
  currentStock: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  in: 'in',
  out: 'out',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  /** @nullable */
  materialName: string | null;
  /** @nullable */
  boxLabel: string | null;
  userName: string;
  count: number;
  createdAt: string;
}

export type ListScanInParams = {
materialId?: number;
};

export type ListScanOutParams = {
materialId?: number;
};

export type ListHistoryParams = {
type?: ListHistoryType;
materialId?: number;
userId?: number;
from?: string;
to?: string;
/**
 * When true, only include scan-in records that still have serial numbers not yet scanned out (excludes scan-out and non-scan rows).
 */
stockOnly?: boolean;
};

export type ListHistoryType = typeof ListHistoryType[keyof typeof ListHistoryType];


export const ListHistoryType = {
  in: 'in',
  out: 'out',
} as const;

export type GetMaterialStatsParams = {
materialId?: number;
};

export type GetRecentActivityParams = {
limit?: number;
};