# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Warehouse Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scan-out:
    get:
      operationId: listScanOut
      tags: [scan-out]
      summary: List all scan-out sessions
      parameters:
        - name: materialId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of scan-out sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanOut"
    post:
      operationId: createScanOut
      tags: [scan-out]
      summary: Create a scan-out session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanOutInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanOut"
  /scan-out/{id}:
    get:
      operationId: getScanOut
      tags: [scan-out]
      summary: Get scan-out session by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ScanOut session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanOut"
        "404":
          description: Not found
    delete:
      operationId: deleteScanOut
      tags: [scan-out]
      summary: Delete a scan-out session (master only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /scan-out/{id}/items:
    post:
      operationId: addScanOutItem
      tags: [scan-out]
      summary: Process a QR code scan for material-out
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanOutItemInput"
      responses:
        "201":
          description: Item processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanItem"
        "404":
          description: Serial number not found in stock
        "409":
          description: Already scanned out
  /scan-out/{id}/items/bulk:
    post:
      operationId: addScanOutItemsBulk
      tags: [scan-out]
      summary: Mark a batch of serial numbers as scanned out in one call (e.g. bulk dispatch triggered from history). Unlike the single-item endpoint, unknown or already-dispatched serials are skipped and reported back instead of failing the whole batch.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanOutItemsBulkInput"
      responses:
        "201":
          description: Batch processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanOutItemsBulkResult"
...
components:
  schemas:
...
    ScanOut:
      type: object
      required: [id, userId, userName, createdAt, items]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScanItem"
    ScanOutInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
    ScanOutItemInput:
      type: object
      required: [qrData]
      properties:
        qrData:
          type: string
    ScanItem:
      type: object
      required: [id, serialNumber, createdAt]
      properties:
        id:
          type: integer
        serialNumber:
          type: string
        scanInId:
          type: ["integer", "null"]
        scanOutId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    ScanOutItemsBulkInput:
      type: object
      required: [serialNumbers]
      properties:
        serialNumbers:
          type: array
          items:
            type: string
    ScanOutItemsBulkResult:
      type: object
      required: [scannedCount, alreadyOut, notFound]
      properties:
        scannedCount:
          type: integer
        alreadyOut:
          type: array
          items:
            type: string
        notFound:
          type: array
          items:
            type: string
...