# lib/api-zod/src/generated/types/scanOutItemsBulkInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Warehouse Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface ScanOutItemsBulkInput {
  serialNumbers: string[];
}